test: create separate test suite for LSP9 proxy
