+++ package.json
@@ -51,7 +51,8 @@
     "test:lsp6init": "hardhat test --no-compile tests/LSP6KeyManager/LSP6KeyManagerInit.test.ts",
     "test:lsp7": "hardhat test --no-compile tests/LSP7DigitalAsset/*.test.ts tests/LSP7DigitalAsset/extensions/*.test.ts",
     "test:lsp8": "hardhat test --no-compile tests/LSP8IdentifiableDigitalAsset/*.test.ts tests/LSP8IdentifiableDigitalAsset/extensions/*.test.ts",
-    "test:lsp9": "hardhat test --no-compile tests/LSP9Vault/*.test.ts",
+    "test:lsp9": "hardhat test --no-compile tests/LSP9Vault/LSP9Vault.test.ts",
+    "test:lsp9init": "hardhat test --no-compile tests/LSP9Vault/LSP9VaultInit.test.ts",
     "test:lsp11": "hardhat test --no-compile tests/LSP11BasicSocialRecovery/*.test.ts",
     "test:lsp20": "hardhat test --no-compile tests/LSP20CallVerification/LSP6/LSP20WithLSP6.test.ts",
     "test:lsp20init": "hardhat test --no-compile tests/LSP20CallVerification/LSP6/LSP20WithLSP6Init.test.ts",
